+++ lib/pharos/cluster_manager.rb
@@ -57,8 +57,8 @@ module Pharos
     end
 
     def gather_facts
-      apply_phase(Phases::ConnectSSH, config.hosts, parallel: true)
-      apply_phase(Phases::AuthenticateSSH, config.hosts.reject(&:ssh?), parallel: false)
+      apply_phase(Phases::ConnectSSH, config.hosts.reject(&:local?), parallel: true)
+      apply_phase(Phases::AuthenticateSSH, config.hosts.reject(&:connected?), parallel: false)
       apply_phase(Phases::GatherFacts, config.hosts, parallel: true)
       apply_phase(Phases::ConfigureClient, [config.master_host], parallel: false, optional: true)
       apply_phase(Phases::LoadClusterConfiguration, [config.master_host]) if config.master_host.master_sort_score.zero?
