+++ lib/pharos/phases/connect_ssh.rb
@@ -7,7 +7,7 @@ module Pharos
 
       def call
         return if host.local?
-        host.ssh(non_interactive: true)
+        host.transport(non_interactive: true)
       rescue Net::SSH::AuthenticationFailed, Net::SSH::Authentication::KeyManagerError
         logger.error { "Authentication failed for #{host.user}@#{host}" }
       end
