+++ lib/pharos/transport.rb
@@ -4,14 +4,14 @@ module Pharos
   module Transport
     def self.for(host, **options)
       if host.local?
-        Local.new(host, **options)
+        Local.new('localhost', **options)
       else
         opts = {}
-        opts[:keys] = [ssh_key_path] if host.ssh_key_path
+        opts[:keys] = [host.ssh_key_path] if host.ssh_key_path
         opts[:send_env] = [] # override default to not send LC_* envs
         opts[:proxy] = Net::SSH::Proxy::Command.new(host.ssh_proxy_command) if host.ssh_proxy_command
         opts[:bastion] = host.bastion if host.bastion
-        SSH.new(address, user: user, **opts.merge(options)).tap(&:connect)
+        SSH.new(host.address, user: host.user, **opts.merge(options)).tap(&:connect)
       end
     end
   end
